Changes for Add and Delete recordings

+++ db/dataset.py
@@ -321,21 +321,23 @@ def add_recordings(dataset, dataset_id):
     """Adds new recordings to a class in a dataset"""
 
     with db.engine.begin() as connection:
-        result = connection.execute("""SELECT id FROM dataset_class WHERE name = %s and dataset = %s""",
-                                    (dataset["class_name"], dataset_id))
+        result = connection.execute(sqlalchemy.text("""SELECT id FROM dataset_class WHERE name = :name and dataset = :dataset_id"""),
+                                    {"name": dataset["class_name"], "dataset_id": dataset_id})
+        if result.rowcount < 1:
+            raise exceptions.NoDataFoundException("No such class exists.")
         clsid = result.fetchone()
-
         for mbid in dataset["recordings"]:
-            connection.execute("""INSERT INTO dataset_class_member (class, mbid) VALUES (%s,%s)""", (clsid[0],mbid))
+            connection.execute(sqlalchemy.text("""INSERT INTO dataset_class_member (class, mbid) VALUES (:clsid,:mbid)"""), {"clsid": clsid[0], "mbid": mbid})
 
 
 def delete_recordings(dataset, dataset_id):
     """Delete all recordings from the list of IDs"""
 
     with db.engine.begin() as connection:
-        result = connection.execute("""SELECT id FROM dataset_class WHERE name = %s and dataset = %s""",
-                                    (dataset["class_name"], dataset_id))
+        result = connection.execute(sqlalchemy.text("""SELECT id FROM dataset_class WHERE name = :name and dataset = :dataset_id"""),
+                                    {"name": dataset["class_name"], "dataset_id": dataset_id})
         clsid = result.fetchone()
-
+        if not clsid:
+            raise db.exceptions.NoDataFoundException("No such class exists.")
         for mbid in dataset["recordings"]:
             connection.execute("""DELETE FROM dataset_class_member WHERE class = %s and mbid = %s""", (clsid[0], mbid))
